+++ c/src/lib/libcpu/arm/pxa255/ffuart/ffuart.c
@@ -25,7 +25,7 @@ volatile int dbg_dly;
 static int     ffuart_first_open(int major, int minor, void *arg);
 static int     ffuart_last_close(int major, int minor, void *arg);
 static int     ffuart_read(int minor);
-static int     ffuart_write(int minor, const char *buf, int len);
+static ssize_t ffuart_write(int minor, const char *buf, size_t len);
 static void    ffuart_init(int minor);
 static void    ffuart_write_polled(int minor, char c);
 static int     ffuart_set_attributes(int minor, const struct termios *t);
@@ -102,9 +102,9 @@ static int ffuart_read(int minor)
  *
  * return 1 on success, -1 on error
  */
-static int ffuart_write(int minor, const char *buf, int len)
+static ssize_t ffuart_write(int minor, const char *buf, size_t len)
 {
-    int i, x;
+    size_t i, x;
     char c;
     console_tbl *console_entry;
     ffuart_reg_t *ffuart;
